todas las putas funciones con bonus incluido pasado por paco y paco -s, la putisima madre que me parió, por fin ;0

+++ libft/ft_strlcpy.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	len;
+	size_t	i;
+	size_t	len;
 
 	i = 0;
 	len = ft_strlen(src);
